Using "saved image" to restore state of viewport before drawing marks.
Something doesn't quire work, I'll need to work on this a bit more. But the basics are there.

I need to stress that this is (as always) just rewriting of GTK functionality in QT, I'm not creating anything original here (maybe just a tiny, tiny bit, to make the best use of C++ and QT, or to change how the app should work a bit).