Improving code related to ParameterScale
The data type now contains 'initial' field that stores initial value of a SpinBox or Scale.
It is also used to store program's hardwired default value. The default value is now stored together with min/max values of the scale, so all the information about a scale/spinbox is kept in one place.
Some more work is needed to actually use the hardwired value that (in case of SpinBoxes and Scales) is now located in new place (not directly in Parameter but in Parameter::ParameterScale::initial).

Also replacing few magic numbers with constants describing waypoint symbols.