Getting rid of ParameterValueTyped data type
Using SGVariant in places where ParameterValueTyped has been used before. Refactoring code related to the old type.
Making a global change in some places, where a new value is assigned to SGVariant variable: now the assignment is made through constructor.
There are still some places where assignment is done directly to a field of SGVariant, without correctly setting data type of the variant. This still needs to be fixed in future. The good news is that we have an assert() that may help catching incorrect variant values.