First callback moved from LayerTRW to sublayer code
Also creating Tree class and a global g_tree object. In the future the object will be converted into a member of TreeItem, as soon as I figure out what exactly is needed by sublayers to work correctly, and how much needs to stay in Tree class definition.
The Tree object/member is intended to be a replacement for different ways of getting window/panel/viewport in sublayer-related code.

Also doing some small renaming of methods.