Beginning to refactor acquiring process
It seems that acquiring process, using different types of data sources, is a big algorithm that tries to treat all types of different data sources in the same way. This has lead into a tree of "if"s and checks and different paths of execution in acquire() function. I will try to fix this.

For now two acquire processes using babel conversion (from file and from OSM) are rewritten a bit.